refactor: split core.h
