+++ w9_pathfinding/src/include/a_star.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "env.h"
+#include "pf.h"
 
 
 class AStar : public AbsPathFinder {
