+++ w9_pathfinding/src/include/bfs.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "env.h"
+#include "pf.h"
 
 
 class BFS : public AbsPathFinder {
