+++ w9_pathfinding/src/include/bi_a_star.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "env.h"
+#include "pf.h"
 
 
 class BiAStar : public AbsPathFinder {
