+++ w9_pathfinding/src/include/bi_dijkstra.h
@@ -1,6 +1,6 @@
  #pragma once
 
-#include "env.h"
+#include "pf.h"
 
 
 class BiDijkstra : public AbsPathFinder {
