+++ w9_pathfinding/src/include/dfs.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "env.h"
+#include "pf.h"
 
 
 class DFS : public AbsPathFinder {
