+++ w9_pathfinding/src/include/dijkstra.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "env.h"
+#include "pf.h"
 
 
 class Dijkstra : public AbsPathFinder {
