+++ w9_pathfinding/src/include/gbs.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "env.h"
+#include "pf.h"
 
 
 class GBS : public AbsPathFinder {
