+++ w9_pathfinding/src/include/hc_a_star.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "env.h"
+#include "mapf.h"
 #include "space_time_a_star.h"
 #include "resumable_search.h"
 #include "reservation_table.h"
