+++ w9_pathfinding/src/include/ida_star.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "env.h"
+#include "pf.h"
 
 
 class IDAStar : public AbsPathFinder {
