+++ w9_pathfinding/src/include/multi_agent_a_star.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <memory>
-#include "env.h"
+#include "mapf.h"
 #include "resumable_search.h"
 #include "space_time_a_star.h"
