+++ w9_pathfinding/src/include/pf.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "env.h"
+
+class AbsPathFinder {
+    public:
+        AbsPathFinder() {};
+        virtual ~AbsPathFinder() {};
+        virtual Path find_path(int start, int end) = 0;
+};
